# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API cho Superapp dành cho Tài xế công nghệ.\nBao gồm: Auth, GPS Tracking, SOS, Chat, Map Sinh tồn, Mood, Earning, Reminder, Emergency Contact.",
        "title": "Grab Driver Superapp API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/emergency/contacts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách anh em thân thiết",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emergency"
                ],
                "summary": "Danh sách liên hệ khẩn cấp",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_emergency.EmergencyContact"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Thêm anh em thân thiết vào danh sách liên hệ khẩn cấp (tối đa 10)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emergency"
                ],
                "summary": "Thêm liên hệ khẩn cấp",
                "parameters": [
                    {
                        "description": "Thông tin liên hệ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_emergency.AddContactRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_emergency.EmergencyContact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/emergency/contacts/{contactId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Xóa anh em khỏi danh sách liên hệ khẩn cấp",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emergency"
                ],
                "summary": "Xóa liên hệ khẩn cấp",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID của liên hệ cần xóa",
                        "name": "contactId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/mood/posts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách bài viết xả mood, mới nhất trước",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mood"
                ],
                "summary": "Feed bài viết",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Số bài (mặc định 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset phân trang",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_mood.Post"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tài xế viết tâm trạng lên bảng tin, có thể ẩn danh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mood"
                ],
                "summary": "Đăng bài xả mood",
                "parameters": [
                    {
                        "description": "Nội dung bài viết",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_mood.CreatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_mood.Post"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/mood/posts/{postId}/comments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy comment của bài viết, cũ nhất trước",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mood"
                ],
                "summary": "Danh sách comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Số comment (mặc định 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset phân trang",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_mood.Comment"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Thêm comment vào bài viết, có thể ẩn danh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mood"
                ],
                "summary": "Bình luận bài viết",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nội dung comment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_mood.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_mood.Comment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/mood/posts/{postId}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Thả like cho bài viết",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mood"
                ],
                "summary": "Like bài viết",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_modules_emergency.AddContactRequest": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "contact_id": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_emergency.EmergencyContact": {
            "type": "object",
            "properties": {
                "alias": {
                    "description": "biệt danh",
                    "type": "string"
                },
                "contact_id": {
                    "description": "userID của anh em",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_mood.Comment": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "author_name": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_anonymous": {
                    "type": "boolean"
                },
                "post_id": {
                    "type": "string"
                }
            }
        },
        "internal_modules_mood.CreateCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "is_anonymous": {
                    "type": "boolean"
                }
            }
        },
        "internal_modules_mood.CreatePostRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "is_anonymous": {
                    "type": "boolean"
                },
                "mood": {
                    "type": "string"
                }
            }
        },
        "internal_modules_mood.Post": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "author_name": {
                    "description": "hiển thị \"Ẩn danh\" nếu anonymous",
                    "type": "string"
                },
                "comment_count": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_anonymous": {
                    "type": "boolean"
                },
                "like_count": {
                    "type": "integer"
                },
                "mood": {
                    "description": "happy, sad, angry, tired, funny",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Nhập JWT token theo format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}